+++ attoparsec-run/library/Data/Attoparsec/ByteString/Run.hs
@@ -1,6 +1,6 @@
 module Data.Attoparsec.ByteString.Run
   (
-    parseStream, parseAndRestore,
+    parseStream, parseAndRestore, parseOnlyStrict, parseOnlyLazy,
     module Data.Attoparsec.Run,
   )
   where
@@ -8,10 +8,17 @@ module Data.Attoparsec.ByteString.Run
 import Data.Attoparsec.Run
 
 import Control.Monad (unless)
+import Control.Monad.State (evalState)
 import Data.Attoparsec.ByteString (parseWith, Parser)
-import Data.ByteString (ByteString)
-import Data.ByteString (null)
+import Data.ByteString (ByteString, null)
+import qualified Data.ByteString.Lazy as Lazy
 import Prelude (($), (<$>), pure, mempty, Monad, Either)
+
+parseOnlyStrict :: ByteString -> Parser a -> Either ParseError a
+parseOnlyStrict x p = evalState (parseAndRestore inputState p) [x]
+
+parseOnlyLazy :: Lazy.ByteString -> Parser a -> Either ParseError a
+parseOnlyLazy x p = evalState (parseAndRestore inputState p) (Lazy.toChunks x)
 
 parseStream :: Monad m =>
     BufferedInput m ByteString -> Parser a -> m (FinalResult ByteString a)