+++ attoparsec-run/library/Data/Attoparsec/Text/Run.hs
@@ -1,6 +1,6 @@
 module Data.Attoparsec.Text.Run
   (
-    parseStream, parseAndRestore,
+    parseStream, parseAndRestore, parseOnlyStrict, parseOnlyLazy,
     module Data.Attoparsec.Run,
   )
   where
@@ -8,10 +8,17 @@ module Data.Attoparsec.Text.Run
 import Data.Attoparsec.Run
 
 import Control.Monad (unless)
+import Control.Monad.State (evalState)
 import Data.Attoparsec.Text (parseWith, Parser)
-import Data.Text (Text)
-import Data.Text (null)
+import Data.Text (Text, null)
+import qualified Data.Text.Lazy as Lazy
 import Prelude (($), (<$>), pure, mempty, Monad, Either)
+
+parseOnlyStrict :: Text -> Parser a -> Either ParseError a
+parseOnlyStrict x p = evalState (parseAndRestore inputState p) [x]
+
+parseOnlyLazy :: Lazy.Text -> Parser a -> Either ParseError a
+parseOnlyLazy x p = evalState (parseAndRestore inputState p) (Lazy.toChunks x)
 
 parseStream :: Monad m =>
     BufferedInput m Text -> Parser a -> m (FinalResult Text a)